Move pages to own directory
